+++ src/apu/channels/common/length_counter.rs
