+++ src/apu/channels/common/timer.rs
